Move find_parent_dossier in og.dossier.utils
