+++ opengever/base/browser/edit_public_trial.py
@@ -1,7 +1,7 @@
 from five import grok
 from opengever.base import _
 from opengever.base.behaviors.classification import IClassification
-from opengever.base.utils import find_parent_dossier
+from opengever.dossier.utils import find_parent_dossier
 from opengever.document.behaviors import IBaseDocument
 from opengever.dossier.base import DOSSIER_STATES_OPEN
 from opengever.dossier.templatefolder.interfaces import ITemplateFolder
