+++ opengever/base/tests/test_utils.py
@@ -4,7 +4,7 @@ from ftw.testing import MockTestCase
 from mocker import ANY
 from opengever.base.behaviors.utils import set_attachment_content_disposition
 from opengever.base.utils import escape_html
-from opengever.base.utils import find_parent_dossier
+from opengever.dossier.utils import find_parent_dossier
 from opengever.testing import FunctionalTestCase
 from plone.namedfile.file import NamedFile
 from unittest2 import TestCase
