+++ opengever/base/utils.py
@@ -1,9 +1,5 @@
-from Acquisition import aq_inner
-from Acquisition import aq_parent
-from opengever.dossier.interfaces import IDossierMarker
 from plone import api
 from Products.CMFCore.utils import getToolByName
-from Products.CMFPlone.interfaces.siteroot import IPloneSiteRoot
 from xml.sax.saxutils import escape
 from zope.component import getMultiAdapter
 import json
@@ -40,21 +36,6 @@ def set_profile_version(portal, profile_id, version):
     assert(ps.getLastVersionForProfile(profile_id) == (version, ))
     print "Set version for '%s' to '%s'." % (profile_id, version)
     return [version]
-
-
-def find_parent_dossier(content):
-    """Returns the first parent dossier relative to the current context.
-    """
-
-    if IPloneSiteRoot.providedBy(content):
-        raise ValueError('Site root passed as argument.')
-
-    while content and not IDossierMarker.providedBy(content):
-        content = aq_parent(aq_inner(content))
-        if IPloneSiteRoot.providedBy(content):
-            raise ValueError('Site root reached while searching '
-                             'parent dossier.')
-    return content
 
 
 def ok_response(request=None):
